Got the shape analysis code to build+run w/ new types.
I had to hack the shit out of it to make it work. There are a lot of things
stubbed out and there is one more large-scale refactoring that still needs to
happen. Specifically the xray-image needs to be threaded throughout the
entire computation so that we can actually allocate new xrays instead of just
pretending that we are.